refactor(db) 2/13: extract db/util.rs shared helpers (closes #248)
Move 14 cross-domain / generic free fns (byte+string+FQN helpers, the
item-object walker, fetch_fqn_payloads, spn-triple parsers, combat-style
maps) into db/util.rs as pub(crate), re-exported via `pub(crate) use
util::*`. Widen Database.conn + with_tx to pub(crate) for sibling domain
modules. Domain-specific classifiers stay until their domain is extracted.
Pure relocation; 239 tests + clippy -D warnings + fmt green.
